+++ src/main/java/org/opentestsystem/ap/migration/migration/Migration2957.java
@@ -1,25 +1,11 @@
 package org.opentestsystem.ap.migration.migration;
 
-import lombok.Data;
-import org.apache.commons.lang3.StringUtils;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
 import org.opentestsystem.ap.common.datastore.DataStoreAttachmentManager;
 import org.opentestsystem.ap.common.datastore.DataStoreDataManager;
 import org.opentestsystem.ap.common.datastore.DataStoreUtility;
 import org.opentestsystem.ap.common.management.ItemManagerEventProducer;
-import org.opentestsystem.ap.common.model.EbsrItem;
-import org.opentestsystem.ap.common.model.EqItem;
 import org.opentestsystem.ap.common.model.Item;
 import org.opentestsystem.ap.common.model.ItemConstants;
-import org.opentestsystem.ap.common.model.ItemImageResource;
-import org.opentestsystem.ap.common.model.SaItem;
-import org.opentestsystem.ap.common.model.WerItem;
-import org.opentestsystem.ap.common.model.content.EbsrItemContent;
-import org.opentestsystem.ap.common.model.content.EqItemContent;
-import org.opentestsystem.ap.common.model.content.SaItemContent;
-import org.opentestsystem.ap.common.model.content.WerItemContent;
 import org.opentestsystem.ap.common.saaif.item.ItemRelease;
 import org.opentestsystem.ap.common.saaif.item.SaaifContent;
 import org.opentestsystem.ap.common.saaif.model.ImportItem;
@@ -27,33 +13,34 @@ import org.opentestsystem.ap.common.saaif.model.SkipMigration;
 import org.opentestsystem.ap.migration.ApplicationDependencyProvider;
 import org.opentestsystem.ap.migration.ApplicationProperties;
 import org.opentestsystem.ap.migration.model.ItemMerge;
+import org.opentestsystem.ap.migration.ttsmapper.TtsMapperManager;
+import org.opentestsystem.ap.migration.ttsmapper.ttsElementInfo;
 import org.springframework.stereotype.Component;
 
-import javax.xml.bind.JAXBElement;
-import java.io.Serializable;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
-
-import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 @Component
 public class Migration2957 extends AbstractImportMigration {
 
     private static String CONTENTLINKINFO_TYPE_EQUATION = "Equation";
 
+    private final TtsMapperManager ttsMapperManager;
+
     public Migration2957(ApplicationProperties applicationProperties,
                          DataStoreDataManager dataManager,
                          ItemManagerEventProducer eventProducer,
                          DataStoreUtility dataStoreUtility,
                          DataStoreAttachmentManager dataStoreAttachmentManager,
-                         ApplicationDependencyProvider applicationDependencyProvider) {
+                         ApplicationDependencyProvider applicationDependencyProvider,
+                         TtsMapperManager ttsMapperManager) {
         super(applicationProperties, dataManager, eventProducer, dataStoreUtility,
                 dataStoreAttachmentManager, applicationDependencyProvider);
+        this.ttsMapperManager = ttsMapperManager;
     }
 
     @Override
@@ -65,7 +52,7 @@ public class Migration2957 extends AbstractImportMigration {
     protected ItemMerge mergeItemFromImportData(ItemMerge itemMerge, ImportItem importItem, Path itemSyncDir) {
         // Set Equation TTS on image resources within the rich text of each item
         mergeEquationTTS(itemMerge.getMergedItem(),
-                getEquationAccessElements(importItem.getItemRelease()));
+                getTtsElementInfoList(importItem.getItemRelease()));
 
         return new ItemMerge(itemMerge.getMergedItem(), itemSyncDir, DO_NOT_SYNC_ATTACHMENTS);
     }
@@ -89,20 +76,24 @@ public class Migration2957 extends AbstractImportMigration {
     }
 
 
-    private void mergeEquationTTS(Item item, List<accessElementInfo> accessElementInfoList) {
+    private void mergeEquationTTS(Item item, List<ttsElementInfo> ttsElementInfoList) {
         String itemType = item.getType();
         switch (itemType) {
             case ItemConstants.ItemType.TYPE_EBSR:
-                mergeEbsrEquationTTS(item, accessElementInfoList);
+                ttsMapperManager.getEbsrTtsMapper()
+                        .mergeEquationTTS(item, ttsElementInfoList);
                 break;
             case ItemConstants.ItemType.TYPE_EQ:
-                mergeEqEquationTTS(item, accessElementInfoList);
+                ttsMapperManager.getEqTtsMapper()
+                        .mergeEquationTTS(item, ttsElementInfoList);
                 break;
             case ItemConstants.ItemType.TYPE_SA:
-                mergeSaEquationTTS(item, accessElementInfoList);
+                ttsMapperManager.getSaTtsMapper()
+                        .mergeEquationTTS(item, ttsElementInfoList);
                 break;
             case ItemConstants.ItemType.TYPE_WER:
-                mergeWerEquationTTS(item, accessElementInfoList);
+                ttsMapperManager.getWerTtsMapper()
+                        .mergeEquationTTS(item, ttsElementInfoList);
                 break;
             default:
                 break;
@@ -110,254 +101,54 @@ public class Migration2957 extends AbstractImportMigration {
     }
 
     // ------------------------------------------------------------------------
-    // Main Rich Text update methods
-
-    private String updateTTSOnEquationImages(String htmlContent, Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        Document document = newDocument(htmlContent);
-        accessElementInfoList.forEach(accessElementInfo -> {
-            Optional<ItemImageResource> imageResource = dataStoreItem.getImages()
-                    .getImageResources().stream()
-                    .filter(resource -> resource.getProductionFile()
-                            .getFileName().equalsIgnoreCase(accessElementInfo.getFileName()))
-                    .findFirst();
-            if (imageResource.isPresent()) {
-                String imageResourceId = imageResource.get().getId();
-                Element image = document.getElementsByAttributeValue("data-iat-image-resource-id", imageResourceId).first();
-                if (image != null) {
-                    Element ttsElement = image.parent();
-                    if (!ttsElement.nodeName().equalsIgnoreCase("span")) {
-                        image.wrap("<span></span>");
-                        ttsElement = image.parent();
-                    }
-                    updateTTSTag(ttsElement, accessElementInfo.getAudioShortDesc());
-                }
-            }
-        });
-        return document.body().html();
-    }
-
-    // ------------------------------------------------------------------------
-    // EbsrItem methods
-    private void mergeEbsrEquationTTS(Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        EbsrItem ebsrDataStoreItem = (EbsrItem) dataStoreItem;
-
-        setEbsrEquationTTS(ebsrDataStoreItem.getCore().getEn(), dataStoreItem, accessElementInfoList);
-        setEbsrEquationTTS(ebsrDataStoreItem.getTranslations().getEsp(), dataStoreItem, accessElementInfoList);
-    }
-
-    private void setEbsrEquationTTS(EbsrItemContent itemContent, Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        itemContent.setPrompt(updateTTSOnEquationImages(
-                itemContent.getPrompt(),
-                dataStoreItem,
-                accessElementInfoList));
-
-        itemContent.setPartAPrompt(updateTTSOnEquationImages(
-                itemContent.getPartAPrompt(),
-                dataStoreItem,
-                accessElementInfoList
-        ));
-
-        itemContent.getPartAOptions().forEach(part -> {
-            part.setText(updateTTSOnEquationImages(
-                    part.getText(),
-                    dataStoreItem,
-                    accessElementInfoList
-            ));
-            part.setRationale(updateTTSOnEquationImages(
-                    part.getRationale(),
-                    dataStoreItem,
-                    accessElementInfoList
-            ));
-        });
-
-        itemContent.setPartBPrompt(updateTTSOnEquationImages(
-                itemContent.getPartBPrompt(),
-                dataStoreItem,
-                accessElementInfoList
-        ));
-
-        itemContent.getPartBOptions().forEach(part -> {
-            part.setText(updateTTSOnEquationImages(
-                    part.getText(),
-                    dataStoreItem,
-                    accessElementInfoList
-            ));
-            part.setRationale(updateTTSOnEquationImages(
-                    part.getRationale(),
-                    dataStoreItem,
-                    accessElementInfoList
-            ));
-        });
-    }
-
-
-    private void updateFileNamesFromEbsrItem(ItemRelease.Item item, List<accessElementInfo> accessElementInfoList){
-        item.getContent().forEach(content -> {
-            // Look for filenames on Stem
-            setFileNamesOnAccessElementInfoList(content.getStem(), accessElementInfoList);
-            // Look for filenames on Qti
-            setFileNamesOnAccessElementInfoList(content.getQti().getValue(), accessElementInfoList);
-        });
-    }
-
-
-    // ------------------------------------------------------------------------
-    // EqItem methods
-
-    private void updateFileNamesFromEqItem(ItemRelease.Item item, List<accessElementInfo> accessElementInfoList) {
-        item.getContent().forEach(content -> {
-            // Look for filenames on Stem
-            setFileNamesOnAccessElementInfoList(content.getStem(), accessElementInfoList);
-            // Look for filenames on Rubric list
-            for (Serializable element : content.getRubriclist().getContent()) {
-                JAXBElement jaxbElement = (JAXBElement) element;
-                if (jaxbElement.getName().toString().equalsIgnoreCase("rubric")) {
-                    ItemRelease.Item.Content.Rubriclist.Rubric rubric =
-                            (ItemRelease.Item.Content.Rubriclist.Rubric) jaxbElement.getValue();
-                    setFileNamesOnAccessElementInfoList(rubric.getVal(), accessElementInfoList);
-                }
-            }
-        });
-    }
-
-    private void mergeEqEquationTTS(Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        EqItem eqDataStoreItem = (EqItem) dataStoreItem;
-
-        setEqEquationTTS(eqDataStoreItem.getCore().getEn(), dataStoreItem, accessElementInfoList);
-        setEqEquationTTS(eqDataStoreItem.getTranslations().getEsp(), dataStoreItem, accessElementInfoList);
-    }
-
-    private void setEqEquationTTS(EqItemContent itemContent, Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        itemContent.setPrompt(updateTTSOnEquationImages(
-                itemContent.getPrompt(),
-                dataStoreItem,
-                accessElementInfoList));
-    }
-
-    // ------------------------------------------------------------------------
-    // SaItem methods
-
-    private void mergeSaEquationTTS(Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        SaItem saDataStoreItem = (SaItem) dataStoreItem;
-
-        setSaEquationTTS(saDataStoreItem.getCore().getEn(), dataStoreItem, accessElementInfoList);
-        setSaEquationTTS(saDataStoreItem.getTranslations().getEsp(), dataStoreItem, accessElementInfoList);
-    }
-
-    private void setSaEquationTTS(SaItemContent itemContent, Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        itemContent.setPrompt(updateTTSOnEquationImages(
-                itemContent.getPrompt(),
-                dataStoreItem,
-                accessElementInfoList));
-
-        itemContent.getRubrics()
-                .forEach(rubric -> rubric = updateTTSOnEquationImages(
-                        rubric, dataStoreItem, accessElementInfoList
-                ));
-
-        itemContent.getExemplarResponses()
-                .forEach(exemplar -> exemplar = updateTTSOnEquationImages(
-                        exemplar, dataStoreItem, accessElementInfoList
-                ));
-    }
-
-    private void updateFileNamesFromSaAndWerItem(ItemRelease.Item item, List<accessElementInfo> accessElementInfoList){
-        item.getContent().forEach(content -> {
-            // Look for filenames on Stem
-            setFileNamesOnAccessElementInfoList(content.getStem(), accessElementInfoList);
-            // Look for filenames on Rubric list
-            for (Serializable element : content.getRubriclist().getContent()) {
-                JAXBElement jaxbElement = (JAXBElement) element;
-                if (jaxbElement.getName().toString().equalsIgnoreCase("rubric")) {
-                    ItemRelease.Item.Content.Rubriclist.Rubric rubric =
-                            (ItemRelease.Item.Content.Rubriclist.Rubric) jaxbElement.getValue();
-                    setFileNamesOnAccessElementInfoList(rubric.getVal(), accessElementInfoList);
-                } else if (jaxbElement.getName().toString().equals("samplelist")) {
-                    ItemRelease.Item.Content.Rubriclist.Samplelist samplelist =
-                            (ItemRelease.Item.Content.Rubriclist.Samplelist) jaxbElement.getValue();
-                    for (ItemRelease.Item.Content.Rubriclist.Samplelist.Sample sample : samplelist.getSample()) {
-                        if (StringUtils.isNotBlank(sample.getSamplecontent())) {
-                            setFileNamesOnAccessElementInfoList(sample.getSamplecontent(), accessElementInfoList);
-                        }
-                    }
-                }
-            }
-        });
-    }
-
-    // ------------------------------------------------------------------------
-    // WerItem methods
-
-    private void mergeWerEquationTTS(Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        WerItem werDataStoreItem = (WerItem) dataStoreItem;
-
-        setWerEquationTTS(werDataStoreItem.getCore().getEn(), dataStoreItem, accessElementInfoList);
-        setWerEquationTTS(werDataStoreItem.getTranslations().getEsp(), dataStoreItem, accessElementInfoList);
-    }
-
-    private void setWerEquationTTS(WerItemContent itemContent, Item dataStoreItem, List<accessElementInfo> accessElementInfoList) {
-        itemContent.setPrompt(updateTTSOnEquationImages(
-                itemContent.getPrompt(),
-                dataStoreItem,
-                accessElementInfoList));
-
-        itemContent.setRubric(updateTTSOnEquationImages(
-                itemContent.getRubric(),
-                dataStoreItem,
-                accessElementInfoList
-        ));
-
-        itemContent.getExemplarResponses()
-                .forEach(exemplar -> exemplar = updateTTSOnEquationImages(
-                        exemplar, dataStoreItem, accessElementInfoList
-                ));
-    }
-
-
-    // ------------------------------------------------------------------------
-    // Methods used to load a list of accessElementInfo objects which contain
+    // Methods used to load a list of ttsElementInfo objects which contain
     // the information necessary to set the TSS value on images
 
-    private List<accessElementInfo> getEquationAccessElements(ItemRelease release) {
+    private List<ttsElementInfo> getTtsElementInfoList(ItemRelease release) {
         if (release.getItem().getFormat().equalsIgnoreCase(ItemConstants.ItemType.TYPE_STIM)) {
-            return loadPassageEquationAccessElements(release.getPassage());
+            return loadPassageTtsElementInfoList(release.getPassage());
         } else {
-            return loadItemEquationAccessElements(release.getItem());
+            return loadItemTtsElementInfoList(release.getItem());
         }
     }
 
-    private List<accessElementInfo> loadPassageEquationAccessElements(ItemRelease.Passage passage) {
+    private List<ttsElementInfo> loadPassageTtsElementInfoList(ItemRelease.Passage passage) {
         return null;
     }
 
-    private List<accessElementInfo> loadItemEquationAccessElements(ItemRelease.Item item) {
-        List<accessElementInfo> accessElementInfoList = new ArrayList<>();
+    private List<ttsElementInfo> loadItemTtsElementInfoList(ItemRelease.Item item) {
+        List<ttsElementInfo> ttsElementInfoList = new ArrayList<>();
         // Collect information from ItemRelease accessibilityInfo elements
         item.getContent().forEach(content -> content.getApipAccessibility().getAccessibilityInfo()
                 .getAccessElement().forEach(accessElement -> {
                     if (accessElement.getContentLinkInfo().getType().equalsIgnoreCase(CONTENTLINKINFO_TYPE_EQUATION)) {
-                        accessElementInfoList.add(new accessElementInfo(content.getLanguage(),
+                        ttsElementInfoList.add(new ttsElementInfo(content.getLanguage(),
                                 accessElement.getContentLinkInfo().getItsLinkIdentifierRef(),
                                 accessElement.getRelatedElementInfo().getReadAloud().getAudioShortDesc()));
                     }
                 }));
         // Set filenames on information collected previously (item type specific)
-        updateFileNamesOnAccessElementInfoList(item, accessElementInfoList);
-        return accessElementInfoList;
+        updateFileNamesOnTtsElementInfoList(item, ttsElementInfoList);
+        return ttsElementInfoList;
     }
 
-    private void updateFileNamesOnAccessElementInfoList(ItemRelease.Item item, List<accessElementInfo> accessElementInfoList) {
+    private void updateFileNamesOnTtsElementInfoList(ItemRelease.Item item, List<ttsElementInfo> ttsElementInfoList) {
         switch (item.getFormat()) {
             case ItemConstants.ItemType.TYPE_EBSR:
-                updateFileNamesFromEbsrItem(item, accessElementInfoList);
+                ttsMapperManager.getEbsrTtsMapper()
+                        .updateFileNamesFromItem(item, ttsElementInfoList);
                 break;
             case ItemConstants.ItemType.TYPE_EQ:
-                updateFileNamesFromEqItem(item, accessElementInfoList);
+                ttsMapperManager.getEqTtsMapper()
+                        .updateFileNamesFromItem(item, ttsElementInfoList);
+                break;
             case ItemConstants.ItemType.TYPE_SA:
-                updateFileNamesFromSaAndWerItem(item, accessElementInfoList);
+                ttsMapperManager.getSaTtsMapper()
+                        .updateFileNamesFromItem(item, ttsElementInfoList);
                 break;
             case ItemConstants.ItemType.TYPE_WER:
-                updateFileNamesFromSaAndWerItem(item, accessElementInfoList);
+                ttsMapperManager.getWerTtsMapper()
+                        .updateFileNamesFromItem(item, ttsElementInfoList);
                 break;
             default:
                 break;
@@ -365,17 +156,6 @@ public class Migration2957 extends AbstractImportMigration {
 
         // TODO: Apply item specific logic to read filenames from Rich Text
     }
-
-    private void setFileNamesOnAccessElementInfoList(String richText, List<accessElementInfo> accessElementInfoList) {
-        Document stemDoc = newDocument(richText);
-        accessElementInfoList.forEach(accessElementInfo -> {
-            Element image = stemDoc.getElementById(accessElementInfo.getItsLinkIdentifierRef());
-            if (Objects.nonNull(image)) {
-                accessElementInfo.setFileName(image.attr("src"));
-            }
-        });
-    }
-
 
     // ------------------------------------------------------------------------
     // Used in checkSkipMigration. Determine is item has Equation accessibilityInfo.accessElement (TTS)
@@ -422,38 +202,5 @@ public class Migration2957 extends AbstractImportMigration {
                             .getType().equalsIgnoreCase(CONTENTLINKINFO_TYPE_EQUATION));
         }
         return hasEquation;
-    }
-
-    // ------------------------------------------------------------------------
-    // Helper methods
-
-    private void updateTTSTag(Element ttsTag, String ttsValue) {
-        if (!ttsTag.hasClass("iat-text2speech")) {
-            ttsTag.addClass("iat-text2speech");
-        }
-        if (isNotBlank(ttsValue)) {
-            ttsTag.attr("data-iat-tts", ttsValue);
-        }
-    }
-
-    private Document newDocument(String modelContent) {
-        Document doc = Jsoup.parseBodyFragment(modelContent);
-        doc.outputSettings(new Document.OutputSettings().prettyPrint(false));
-        doc.outputSettings().syntax(Document.OutputSettings.Syntax.xml);
-        return doc;
-    }
-
-    @Data
-    private class accessElementInfo {
-        private String language;
-        private String itsLinkIdentifierRef;
-        private String audioShortDesc;
-        private String fileName;
-
-        accessElementInfo(String language, String itsLinkIdentifierRef, String audioShortDesc) {
-            this.language = language;
-            this.itsLinkIdentifierRef = itsLinkIdentifierRef;
-            this.audioShortDesc = audioShortDesc;
-        }
     }
 }
