+++ src/main/java/org/opentestsystem/ap/migration/ttsmapper/EqTtsMapper.java
@@ -1,0 +1,46 @@
+package org.opentestsystem.ap.migration.ttsmapper;
+
+import org.opentestsystem.ap.common.model.EqItem;
+import org.opentestsystem.ap.common.model.Item;
+import org.opentestsystem.ap.common.model.content.EqItemContent;
+import org.opentestsystem.ap.common.saaif.item.ItemRelease;
+import org.springframework.stereotype.Component;
+
+import javax.xml.bind.JAXBElement;
+import java.io.Serializable;
+import java.util.List;
+
+@Component
+public class EqTtsMapper extends TtsMapper {
+    @Override
+    public void mergeEquationTTS(Item dataStoreItem, List<ttsElementInfo> ttsElementInfoList) {
+        EqItem eqDataStoreItem = (EqItem) dataStoreItem;
+
+        setEquationTTS(eqDataStoreItem.getCore().getEn(), dataStoreItem, ttsElementInfoList);
+        setEquationTTS(eqDataStoreItem.getTranslations().getEsp(), dataStoreItem, ttsElementInfoList);
+    }
+
+    @Override
+    public void updateFileNamesFromItem(ItemRelease.Item item, List<ttsElementInfo> ttsElementInfoList) {
+        item.getContent().forEach(content -> {
+            // Look for filenames on Stem
+            setFileNamesOnAccessElementInfoList(content.getStem(), ttsElementInfoList);
+            // Look for filenames on Rubric list
+            for (Serializable element : content.getRubriclist().getContent()) {
+                JAXBElement jaxbElement = (JAXBElement) element;
+                if (jaxbElement.getName().toString().equalsIgnoreCase("rubric")) {
+                    ItemRelease.Item.Content.Rubriclist.Rubric rubric =
+                            (ItemRelease.Item.Content.Rubriclist.Rubric) jaxbElement.getValue();
+                    setFileNamesOnAccessElementInfoList(rubric.getVal(), ttsElementInfoList);
+                }
+            }
+        });
+    }
+
+    private void setEquationTTS(EqItemContent itemContent, Item dataStoreItem, List<ttsElementInfo> ttsElementInfoList) {
+        itemContent.setPrompt(updateTTSOnEquationImages(
+                itemContent.getPrompt(),
+                dataStoreItem,
+                ttsElementInfoList));
+    }
+}
