+++ src/main/java/org/opentestsystem/ap/migration/ttsmapper/SaTtsMapper.java
@@ -1,0 +1,65 @@
+package org.opentestsystem.ap.migration.ttsmapper;
+
+import org.apache.commons.lang3.StringUtils;
+import org.opentestsystem.ap.common.model.Item;
+import org.opentestsystem.ap.common.model.SaItem;
+import org.opentestsystem.ap.common.model.content.SaItemContent;
+import org.opentestsystem.ap.common.saaif.item.ItemRelease;
+import org.springframework.stereotype.Component;
+
+import javax.xml.bind.JAXBElement;
+import java.io.Serializable;
+import java.util.List;
+
+@Component
+public class SaTtsMapper extends TtsMapper {
+    @Override
+    public void mergeEquationTTS(Item dataStoreItem, List<ttsElementInfo> ttsElementInfoList) {
+        SaItem saDataStoreItem = (SaItem) dataStoreItem;
+
+        setEquationTTS(saDataStoreItem.getCore().getEn(), dataStoreItem, ttsElementInfoList);
+        setEquationTTS(saDataStoreItem.getTranslations().getEsp(), dataStoreItem, ttsElementInfoList);
+    }
+
+    @Override
+    public void updateFileNamesFromItem(ItemRelease.Item item, List<ttsElementInfo> ttsElementInfoList) {
+        item.getContent().forEach(content -> {
+            // Look for filenames on Stem
+            setFileNamesOnAccessElementInfoList(content.getStem(), ttsElementInfoList);
+            // Look for filenames on Rubric list
+            for (Serializable element : content.getRubriclist().getContent()) {
+                JAXBElement jaxbElement = (JAXBElement) element;
+                if (jaxbElement.getName().toString().equalsIgnoreCase("rubric")) {
+                    ItemRelease.Item.Content.Rubriclist.Rubric rubric =
+                            (ItemRelease.Item.Content.Rubriclist.Rubric) jaxbElement.getValue();
+                    setFileNamesOnAccessElementInfoList(rubric.getVal(), ttsElementInfoList);
+                } else if (jaxbElement.getName().toString().equals("samplelist")) {
+                    ItemRelease.Item.Content.Rubriclist.Samplelist samplelist =
+                            (ItemRelease.Item.Content.Rubriclist.Samplelist) jaxbElement.getValue();
+                    for (ItemRelease.Item.Content.Rubriclist.Samplelist.Sample sample : samplelist.getSample()) {
+                        if (StringUtils.isNotBlank(sample.getSamplecontent())) {
+                            setFileNamesOnAccessElementInfoList(sample.getSamplecontent(), ttsElementInfoList);
+                        }
+                    }
+                }
+            }
+        });
+    }
+
+    private void setEquationTTS(SaItemContent itemContent, Item dataStoreItem, List<ttsElementInfo> ttsElementInfoList) {
+        itemContent.setPrompt(updateTTSOnEquationImages(
+                itemContent.getPrompt(),
+                dataStoreItem,
+                ttsElementInfoList));
+
+        itemContent.getRubrics()
+                .forEach(rubric -> rubric = updateTTSOnEquationImages(
+                        rubric, dataStoreItem, ttsElementInfoList
+                ));
+
+        itemContent.getExemplarResponses()
+                .forEach(exemplar -> exemplar = updateTTSOnEquationImages(
+                        exemplar, dataStoreItem, ttsElementInfoList
+                ));
+    }
+}
