+++ src/main/java/org/opentestsystem/ap/migration/ttsmapper/TtsMapperManager.java
@@ -1,0 +1,21 @@
+package org.opentestsystem.ap.migration.ttsmapper;
+
+import lombok.Data;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+public class TtsMapperManager {
+
+    private final EbsrTtsMapper ebsrTtsMapper;
+    private final EqTtsMapper eqTtsMapper;
+    private final SaTtsMapper saTtsMapper;
+    private final WerTtsMapper werTtsMapper;
+
+    public TtsMapperManager(EbsrTtsMapper ebsrTtsMapper, EqTtsMapper eqTtsMapper, SaTtsMapper saTtsMapper, WerTtsMapper werTtsMapper) {
+        this.ebsrTtsMapper = ebsrTtsMapper;
+        this.eqTtsMapper = eqTtsMapper;
+        this.saTtsMapper = saTtsMapper;
+        this.werTtsMapper = werTtsMapper;
+    }
+}
