+++ src/main/java/org/opentestsystem/ap/migration/ttsmapper/WerTtsMapper.java
@@ -1,0 +1,67 @@
+package org.opentestsystem.ap.migration.ttsmapper;
+
+import org.apache.commons.lang3.StringUtils;
+import org.opentestsystem.ap.common.model.Item;
+import org.opentestsystem.ap.common.model.WerItem;
+import org.opentestsystem.ap.common.model.content.WerItemContent;
+import org.opentestsystem.ap.common.saaif.item.ItemRelease;
+import org.springframework.stereotype.Component;
+
+import javax.xml.bind.JAXBElement;
+import java.io.Serializable;
+import java.util.List;
+
+@Component
+public class WerTtsMapper extends TtsMapper {
+
+    @Override
+    public void mergeEquationTTS(Item dataStoreItem, List<ttsElementInfo> ttsElementInfoList) {
+        WerItem werDataStoreItem = (WerItem) dataStoreItem;
+
+        setEquationTTS(werDataStoreItem.getCore().getEn(), dataStoreItem, ttsElementInfoList);
+        setEquationTTS(werDataStoreItem.getTranslations().getEsp(), dataStoreItem, ttsElementInfoList);
+    }
+
+    @Override
+    public void updateFileNamesFromItem(ItemRelease.Item item, List<ttsElementInfo> ttsElementInfoList) {
+        item.getContent().forEach(content -> {
+            // Look for filenames on Stem
+            setFileNamesOnAccessElementInfoList(content.getStem(), ttsElementInfoList);
+            // Look for filenames on Rubric list
+            for (Serializable element : content.getRubriclist().getContent()) {
+                JAXBElement jaxbElement = (JAXBElement) element;
+                if (jaxbElement.getName().toString().equalsIgnoreCase("rubric")) {
+                    ItemRelease.Item.Content.Rubriclist.Rubric rubric =
+                            (ItemRelease.Item.Content.Rubriclist.Rubric) jaxbElement.getValue();
+                    setFileNamesOnAccessElementInfoList(rubric.getVal(), ttsElementInfoList);
+                } else if (jaxbElement.getName().toString().equals("samplelist")) {
+                    ItemRelease.Item.Content.Rubriclist.Samplelist samplelist =
+                            (ItemRelease.Item.Content.Rubriclist.Samplelist) jaxbElement.getValue();
+                    for (ItemRelease.Item.Content.Rubriclist.Samplelist.Sample sample : samplelist.getSample()) {
+                        if (StringUtils.isNotBlank(sample.getSamplecontent())) {
+                            setFileNamesOnAccessElementInfoList(sample.getSamplecontent(), ttsElementInfoList);
+                        }
+                    }
+                }
+            }
+        });
+    }
+
+    private void setEquationTTS(WerItemContent itemContent, Item dataStoreItem, List<ttsElementInfo> ttsElementInfoList) {
+        itemContent.setPrompt(updateTTSOnEquationImages(
+                itemContent.getPrompt(),
+                dataStoreItem,
+                ttsElementInfoList));
+
+        itemContent.setRubric(updateTTSOnEquationImages(
+                itemContent.getRubric(),
+                dataStoreItem,
+                ttsElementInfoList
+        ));
+
+        itemContent.getExemplarResponses()
+                .forEach(exemplar -> exemplar = updateTTSOnEquationImages(
+                        exemplar, dataStoreItem, ttsElementInfoList
+                ));
+    }
+}
