+++ src/main/java/org/opentestsystem/ap/migration/ttsmapper/ttsElementInfo.java
@@ -1,0 +1,17 @@
+package org.opentestsystem.ap.migration.ttsmapper;
+
+import lombok.Data;
+
+@Data
+public class ttsElementInfo {
+    private String language;
+    private String itsLinkIdentifierRef;
+    private String audioShortDesc;
+    private String fileName;
+
+    public ttsElementInfo(String language, String itsLinkIdentifierRef, String audioShortDesc) {
+        this.language = language;
+        this.itsLinkIdentifierRef = itsLinkIdentifierRef;
+        this.audioShortDesc = audioShortDesc;
+    }
+}
